dynaface-lib/requirements.txt
torch>=1.4.0
opencv-python>=4.7.0
Pillow>=8.4.0
matplotlib>=3.7.1
scikit-learn>=1.2.2
plotly>=5.9.0
kaleido>=0.2.1
numpy>=1.18.2
torchvision>=0.5.0
facenet-pytorch
torchaudio
scipy
